BUG Fix interpolation of int images using order=1
A conversion to floating point is necessary

+++ mahotas/interpolate.py
@@ -158,7 +158,7 @@ def _maybe_filter(array, order, func, prefilter, dtype):
     if prefilter and order > 1:
         return spline_filter(array, order, dtype=dtype)
     else:
-        return array
+        return array.astype(dtype)
 
 def zoom(array, zoom, out=None, order=3, mode='constant', cval=0.0, prefilter=True, output=None):
     """
